2011-04-15  Emil A Eklund  <eae@chromium.org>
        Reviewed by Dimitri Glazkov.

        input field with focus makes appendChild operation ~42x slower
        https://bugs.webkit.org/show_bug.cgi?id=57059

        Add performance test for cloneNode when an input field has focus.

        * perf/clone-with-focus-expected.txt: Added.
        * perf/clone-with-focus.html: Added.
2011-04-15  Emil A Eklund  <eae@chromium.org>

        Reviewed by Dimitri Glazkov.

        input field with focus makes appendChild operation ~42x slower
        https://bugs.webkit.org/show_bug.cgi?id=57059

        Change ContainerNode::cloneChildNodes to only disable the
        deleteButtonController if the container itself (or any of its children)
        is being edited. Thus avoiding a reflow in cases where it's not.

        Test: perf/clone-with-focus.html

        * dom/ContainerNode.cpp:
        (WebCore::ContainerNode::cloneChildNodes):

Canonical link: https://commits.webkit.org/73797@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@84056 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/dom/ContainerNode.cpp
@@ -798,9 +798,16 @@ void ContainerNode::childrenChanged(bool changedByParser, Node* beforeChange, No
 void ContainerNode::cloneChildNodes(ContainerNode *clone)
 {
     // disable the delete button so it's elements are not serialized into the markup
-    bool isEditorEnabled = document()->frame() && document()->frame()->editor()->canEdit();
-    if (isEditorEnabled)
-        document()->frame()->editor()->deleteButtonController()->disable();
+    bool isEditorEnabled = false;
+    if (document()->frame() && document()->frame()->editor()->canEdit()) {
+        SelectionController* selection = document()->frame()->selection();
+        Element* root = selection ? selection->rootEditableElement() : 0;
+        isEditorEnabled = root && isDescendantOf(root);
+
+        if (isEditorEnabled)
+            document()->frame()->editor()->deleteButtonController()->disable();
+    }
+    
     ExceptionCode ec = 0;
     for (Node* n = firstChild(); n && !ec; n = n->nextSibling())
         clone->appendChild(n->cloneNode(true), ec);
